NewBook component test
